esut_econs_2011/backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djoser>=2.2
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
pillow>=10.0
